% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goodslides-package.R, R/goodslides.R
\docType{package}
\name{goodslides}
\alias{goodslides}
\alias{goodslide}
\alias{goodslides-package}
\alias{goodslides}
\title{goodslides}
\usage{
goodslides(f, ...)
}
\arguments{
\item{f}{Name of file - must use either .md or .Rmd file extension.}

\item{\dots}{Passed to goodslides.theme.}
}
\value{
Saved html file using same name but with .html extension.
}
\description{
goodslides provides a simple way for good html
  slides in R.

Compiles markdown file to html5 slides
}
\examples{
\dontrun{
## convert slides.md to html slides presentation
goodslides("slides.md")
}

}
