% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{goodslides.theme}
\alias{goodslides.theme}
\title{goodslides.theme}
\usage{
goodslides.theme(base_size = NULL, base_weight = NULL, base_family = NULL,
  base_color = NULL, code_family = NULL, heading_family = NULL,
  heading_weight = NULL, color1 = NULL, color2 = NULL, color3 = NULL,
  logo1 = NULL, logo2 = NULL, logo3 = NULL, stroke = TRUE)
}
\arguments{
\item{base_size}{Size of base font. Defaults to 28.}

\item{base_weight}{Weight of base font. Defaults to 300.}

\item{base_family}{Base font family. Defaults to Roboto.}

\item{base_color}{Color of base font. Defaults to #111.}

\item{code_family}{Code font family. Defaults to Inconsolata.}

\item{heading_family}{Heading font family. Defaults to Montserrat.}

\item{heading_weight}{Weight of heading font. Defaults to 600.}

\item{color1}{Primary theme color. Defaults to yellowgreen.}

\item{color2}{Secondary theme color. Defaults to dark purple.}

\item{color3}{Optional color used for stroke outline.}

\item{logo1}{First logo.}

\item{logo2}{Second logo.}

\item{logo3}{Third logo. For transition slides.}

\item{stroke}{Logical indicating whether to use dark outline on slide headings.
Default to true.}
}
\value{
Saves modified style file.
}
\description{
Sets style theme for html slides.
}
